+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/trie/bonsai/AbstractIsolationTests.java
@@ -207,11 +207,6 @@ public abstract class AbstractIsolationTests {
               public int getDatabaseVersion() {
                 return 2;
               }
-
-              @Override
-              public boolean getDowngradeAllowed() {
-                return false;
-              }
             })
         .withMetricsSystem(new NoOpMetricsSystem())
         .build();