+++ plugin-api/src/main/java/org/hyperledger/besu/plugin/services/BesuConfiguration.java
@@ -44,6 +44,4 @@ public interface BesuConfiguration extends BesuService {
   default int getDatabaseVersion() {
     return 1;
   }
-
-  boolean getDowngradeAllowed();
 }